% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModularIdentifytSNE.R
\name{IdentifytSNEUI}
\alias{IdentifytSNEUI}
\title{IdentifytSNEUI}
\usage{
IdentifytSNEUI(id)
}
\arguments{
\item{id}{The namespace for the module}
}
\description{
Renders tSNE or PCA plot from a seurat object
}
\seealso{
shiny
}
\author{
Kirk Gosik <kgosik@broadinstitute.org>
}
\keyword{shiny}
