% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script}
\alias{script}
\title{Identify Script Name}
\usage{
script()
}
\value{
character: path to current file, or empty string if indeterminate
}
\description{
Identifies script name in a variety of contexts,
e.g. interactively or when script is sourced.
Attempts to support RStudio environment.
}
\examples{
script()
}
