% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{scriptbase}
\alias{scriptbase}
\title{Identify Script base Name}
\usage{
scriptbase(ext = NULL)
}
\arguments{
\item{ext}{replacement extension}
}
\value{
character: current filename, or empty string if indeterminate
}
\description{
Identifies script base name using \code{\link{script}}.
if \code{ext} is supplied (and \code{script() is not empty string}),
the last dot (and anything that follows it) is removed if present;
 \code{ext} is appended.
}
\examples{
scriptbase('.csv')
}
