% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRiptUI.R
\name{scRiptUI}
\alias{scRiptUI}
\title{scRiptUI}
\usage{
scRiptUI()
}
\description{
Launch the single cell R interactive playground tool user interface for your local machine
}
\details{
This function will automatically launch the scRipt user interface in a web browser.
The user interface will allow you to create static reports about your data.  It will also allow you
to interactively explore the data in more detail.
}
\examples{
\dontrun{
   scRiptUI()
}
}
\seealso{
shiny
}
\author{
Kirk Gosik <kgosik@broadinstitute.org>
}
