% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModularUMItSNEPlot.R
\name{UMItSNEPlotServer}
\alias{UMItSNEPlotServer}
\title{UMItSNEPlotServer}
\usage{
UMItSNEPlotServer(input, output, session, outs)
}
\arguments{
\item{input}{List-like object that stores the current values of all of the widgets in your app.}

\item{output}{List-like object that stores instructions for building the R objects in your app.}

\item{session}{List-like object about the session to be passed to the UI.}

\item{outs}{a reactive object from the cellranger toolkit}
}
\description{
Renders tSNE or PCA plot from a Seurat object
}
\seealso{
shiny
}
\author{
Kirk Gosik <kgosik@broadinstitute.org>
}
\keyword{shiny}
